requirements.txt
streamlit>=1.28.0
langchain>=0.1.0
langgraph>=0.0.10
langchain-anthropic>=0.0.10
anthropic>=0.8.1
pydantic>=2.5.2
tiktoken>=0.5.1
requests>=2.31.0
beautifulsoup4>=4.12.2
python-dotenv>=1.0.0
notion-client>=2.0.0
pytest>=7.4.0